+++ tests/testthat/test-deployments-find.R
@@ -1,12 +1,33 @@
-test_that("errors if no deployments", {
+test_that("error when no deployments and no accounts", {
+  local_temp_config()
+
   app <- local_temp_app()
-  expect_snapshot(findDeployment(app), error = TRUE)
+
+  expect_snapshot(findDeployment(app, appName = "placeholder"), error = TRUE)
+})
+
+test_that("returns stubbed details when single account has no deployments", {
+  local_temp_config()
+  addTestServer()
+  addTestAccount()
+
+  app <- local_temp_app()
+
+  # derived from directory name.
+  dep <- findDeployment(app)
+  expect_equal(dep, list(name = basename(app), account = "ron", server = "example.com"))
+
+  # name given.
+  dep <- findDeployment(app, appName = "placeholder")
+  expect_equal(dep, list(name = "placeholder", account = "ron", server = "example.com"))
 })
 
 test_that("finds single deployment", {
-  app <- local_temp_app()
+  local_temp_config()
   addTestServer()
   addTestAccount()
+
+  app <- local_temp_app()
   addTestDeployment(app)
 
   dep <- findDeployment(app)
@@ -14,9 +35,11 @@ test_that("finds single deployment", {
 })
 
 test_that("disambiguates multiple deployments", {
-  app <- local_temp_app()
+  local_temp_config()
   addTestServer()
   addTestAccount()
+
+  app <- local_temp_app()
   addTestDeployment(app, "test1")
   addTestDeployment(app, "test2")
   expect_snapshot(findDeployment(app), error = TRUE)