+++ NEWS.md
@@ -1,5 +1,9 @@
 # rsconnect (development version)
 
+* `showLogs()`, `configureApp()`, `setProperty()`, and `unsetProperty()`
+  search for the application by name when there are no matching deployment
+  records. (#985, #989)
+
 # rsconnect 1.1.1
 
 * Added `space` parameter to deploy directly to a space in Posit Cloud.